requirements.txt
python-pptx==0.6.21
beautifulsoup4==4.12.2 